requirements.txt
django
djangorestframework
markdown
django-filter
django-webpack-loader
Fabric
